+++ app/src/main/kotlin/ac/mdiq/podcini/storage/model/AppPrefs.kt
@@ -55,6 +55,9 @@ class AppPrefs: RealmObject {
     var autoBackupLimit: Int = 2
     var autoBackupTimeStamp: Long = 0L
 
+    var postRepeatsInternal: Int = 24
+    var postRepeatsTime: Long = 0L
+
     var useCustomMediaFolder: Boolean = false
     var customMediaUri: String = ""
 
@@ -158,6 +161,8 @@ class AppPrefs: RealmObject {
         if (autoBackupIntervall != other.autoBackupIntervall) return false
         if (autoBackupLimit != other.autoBackupLimit) return false
         if (autoBackupTimeStamp != other.autoBackupTimeStamp) return false
+        if (postRepeatsInternal != other.postRepeatsInternal) return false
+        if (postRepeatsTime != other.postRepeatsTime) return false
         if (useCustomMediaFolder != other.useCustomMediaFolder) return false
         if (customFolderUnavailable != other.customFolderUnavailable) return false
         if (autoDelete != other.autoDelete) return false
@@ -238,6 +243,8 @@ class AppPrefs: RealmObject {
         result = 31 * result + autoBackupIntervall
         result = 31 * result + autoBackupLimit
         result = 31 * result + autoBackupTimeStamp.hashCode()
+        result = 31 * result + postRepeatsInternal.hashCode()
+        result = 31 * result + postRepeatsTime.hashCode()
         result = 31 * result + useCustomMediaFolder.hashCode()
         result = 31 * result + customFolderUnavailable.hashCode()
         result = 31 * result + autoDelete.hashCode()