+++ app/src/main/kotlin/ac/mdiq/podcini/storage/utils/AutoBackup.kt
@@ -45,7 +45,6 @@ fun autoBackup() {
     }
 
     val curTime = nowInMillis()
-//    if ((curTime - appPrefs.autoBackupTimeStamp) / 10000 > appPrefs.autoBackupIntervall)
     if ((curTime - appPrefs.autoBackupTimeStamp) / 3600000 > appPrefs.autoBackupIntervall)
         CoroutineScope(Dispatchers.IO).launch {
             val uri = uriString.toSafeUri()