+++ app/src/main/kotlin/ac/mdiq/podcini/ui/screens/MainScreen.kt
@@ -81,7 +81,7 @@ var playerMinHeight by mutableIntStateOf(100)
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun MainActivityUI() {
+fun MainScreen() {
     val lifecycleOwner = LocalLifecycleOwner.current
     val context by rememberUpdatedState(LocalContext.current)
     val lcScope = rememberCoroutineScope()